refactor: splits off 'main' code from the cli
