+++ src/run-cli.ts
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+
+import { cli } from "./cli.js";
+
+cli();
