+++ types/cli.d.ts
@@ -1,0 +1,10 @@
+export interface ICommandLineOptions {
+  virtualCodeOwners: string;
+  virtualTeams: string;
+  codeOwners: string;
+  emitLabeler: boolean;
+  labelerLocation: string;
+  dryRun: boolean;
+  help: boolean;
+  version: boolean;
+}
