+++ types/generate-labeler-yml.d.ts
@@ -1,4 +1,5 @@
 import type { ITeamMap, IVirtualCodeOwnersCST } from "./types.js";
+
 export default function generateLabelerYml(
   pCodeOwners: IVirtualCodeOwnersCST,
   pTeamMap: ITeamMap,
